Switched to clangd extension instead of intellisense.

+++ targetloc/targetloc.cpp
@@ -5,18 +5,23 @@ void TargetLoc::start()
 {
     cm.start();
 
-    cameraL.registerCallback([&](const cv::Mat &left, const libcamera::ControlList &)
-                             { updateImageL(left); });
+    cameraL.registerCallback(
+        [&](const cv::Mat &left, const libcamera::ControlList &) {
+            updateImageL(left);
+        });
 
-    cameraR.registerCallback([&](const cv::Mat &right, const libcamera::ControlList &)
-                             { updateImageR(right); });
+    cameraR.registerCallback(
+        [&](const cv::Mat &right, const libcamera::ControlList &) {
+            updateImageR(right);
+        });
 
-    targetDet.registerDetCallback([&](const std::vector<cv::Point2f> &coords)
-                                  { onTargetDetected(coords); });
+    targetDet.registerDetCallback([&](const std::vector<cv::Point2f> &coords) {
+        onTargetDetected(coords);
+    });
 
     stereo.registerCallback([&](cv::Mat d) {
-	std::lock_guard<std::mutex> guard(disparityData_mutex);
-	currentD = d; 
+        std::lock_guard<std::mutex> guard(disparityData_mutex);
+        currentD = d;
     });
 
     settings.width = 1920;
@@ -38,10 +43,8 @@ void TargetLoc::newScanAvail(C1LidarData (&data)[C1Lidar::nDistance])
 {
     std::lock_guard<std::mutex> guard(lidarData_mutex);
     currentLidarCoords.clear();
-    for (const auto &v : data)
-    {
-        if (v.valid)
-        {
+    for (const auto &v : data) {
+        if (v.valid) {
             currentLidarCoords.push_back({v.x, v.y});
         }
     }
@@ -86,19 +89,20 @@ void TargetLoc::onTargetDetected(const std::vector<cv::Point2f> &contour)
     float avgX = 0;
     contour_mutex.lock();
     scaledContour.clear();
-    for(auto &c:contour) {
-	const int x = c.x * currentD.size().width / settings.width;
-	const int y = c.y * currentD.size().height / settings.height;
-	scaledContour.emplace_back(x,y);
-	printf("[%d,%d]", x, y);
-	avgX = avgX + c.x;
-	i++;
+    for (auto &c : contour) {
+        const int x = c.x * currentD.size().width / settings.width;
+        const int y = c.y * currentD.size().height / settings.height;
+        scaledContour.emplace_back(x, y);
+        printf("[%d,%d]", x, y);
+        avgX = avgX + c.x;
+        i++;
     }
-    contour_mutex.unlock();    
+    contour_mutex.unlock();
     avgX = avgX / i;
     printf("\n");
 
-    if (currentD.empty()) return;
+    if (currentD.empty())
+        return;
 
     // Create mask
     cv::Mat mask = cv::Mat::zeros(currentD.size(), CV_8UC1);
@@ -106,19 +110,19 @@ void TargetLoc::onTargetDetected(const std::vector<cv::Point2f> &contour)
 
     // Draw filled contour for the mask
     std::vector<std::vector<cv::Point>> scaledContours{scaledContour};
-    cv::drawContours(mask, scaledContours,-1, cv::Scalar(255), cv::FILLED);
+    cv::drawContours(mask, scaledContours, -1, cv::Scalar(255), cv::FILLED);
 
     // Compute mean gray value inside contour
     float avgDisp = cv::mean(currentD, mask)[0];
 
     const float targetLocX = disp2meter / avgDisp;
-    const float targetLocY = xpos2meter * ((settings.width/2) - avgX);
+    const float targetLocY = xpos2meter * ((int)(settings.width / 2) - avgX);
 
-    printf("Disparity: %f, Target location: [%f,%f]\n",avgDisp,targetLocX,targetLocY);
+    printf("Disparity: %f, Target location: [%f,%f]\n", avgDisp, targetLocX,
+           targetLocY);
 
     // callback!
-    if (detectionInterface)
-    {
+    if (detectionInterface) {
         detectionInterface->newTargetDetected(targetLocX, targetLocY);
     }
 }
